# frontend/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenApi specifiction - LEARNCORE",
    "description": "OpenApi documentation for LEARNCORE project",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "jotsamikael",
      "url": "https://dodibo.com/",
      "email": "jotsamikael@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://someurl.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV."
    },
    {
      "url": "https://something:8088/api/v1",
      "description": "Prod ENV."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/platform/tenants": {
      "get": {
        "tags": [
          "Platform Tenants"
        ],
        "operationId": "listTenants",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Platform Tenants"
        ],
        "operationId": "createTenant",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/staff": {
      "get": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "listPlatformStaff",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "createPlatformStaff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/staff/tenant-admin": {
      "post": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "createTenantAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/staff": {
      "get": {
        "tags": [
          "Tenant Staff"
        ],
        "operationId": "listTenantStaff",
        "parameters": [
          {
            "name": "tenantUuid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tenant Staff"
        ],
        "operationId": "createTenantStaff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/platform/tenants/{uuid}": {
      "get": {
        "tags": [
          "Platform Tenants"
        ],
        "operationId": "getTenant",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Platform Tenants"
        ],
        "operationId": "updateTenant",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PlatformUpdateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/tenants/{uuid}/status": {
      "patch": {
        "tags": [
          "Platform Tenants"
        ],
        "operationId": "changeStatus",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/staff/{uuid}": {
      "get": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "getPlatformStaffDetails",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStaffDetailsResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "updatePlatformStaff",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlatformStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/platform/staff/{uuid}/status": {
      "patch": {
        "tags": [
          "Platform Staff"
        ],
        "operationId": "changeStaffStatus",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tenant": {
      "get": {
        "tags": [
          "Tenant Admin Settings"
        ],
        "operationId": "getOwnTenant",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tenant Admin Settings"
        ],
        "operationId": "updateOwnTenant",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/students/{uuid}/status": {
      "patch": {
        "tags": [
          "Students"
        ],
        "operationId": "activateDeactivate",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/staff/{uuid}": {
      "get": {
        "tags": [
          "Tenant Staff"
        ],
        "operationId": "getTenantStaffDetails",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantStaffDetailsResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tenant Staff"
        ],
        "operationId": "updateTenantStaff",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/staff/{uuid}/status": {
      "patch": {
        "tags": [
          "Tenant Staff"
        ],
        "operationId": "changeStaffStatus_1",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "getStudents",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FindStudentsQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageStudentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/students/{uuid}": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "getStudentDetails",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentDetailResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTenantRequest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "examFocus": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 150,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "TenantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "uuid": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "examFocus": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "CreatePlatformStaffRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "minLength": 1
          },
          "lastname": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1
          },
          "positionName": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstname",
          "lastname",
          "roles"
        ]
      },
      "StaffResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "tenantUuid": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "platformStaff": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "CreateTenantStaffRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "minLength": 1
          },
          "lastname": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1
          },
          "positionName": {
            "type": "string"
          },
          "tenantUuid": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstname",
          "lastname",
          "roles"
        ]
      },
...
      "PlatformUpdateTenantRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "examFocus": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "logoUrl": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 150,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdatePlatformStaffRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "lastname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "positionName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateTenantRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "examFocus": {
            "type": "string",
            "maxLength": 50,
            "minLength": 0
          },
          "logoUrl": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 150,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 30,
            "minLength": 0
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTenantStaffRequest": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "lastname": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "positionName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PlatformStaffDetailsResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatarUrl": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FindStudentsQuery": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sortBy": {
            "type": "string"
          },
          "sortDirection": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "xp": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "streakDays": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "tenantUuid": {
            "type": "string"
          }
        }
      },
      "PageStudentResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "TenantStaffDetailsResponse": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "tenantUuid": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avatarUrl": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "accountLocked": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}